revert
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@56598 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-6.1-nss/drivers/clk/qcom/clk-rcg.h
@@ -14,7 +14,7 @@
 
 struct freq_conf {
 	u8 src;
-	u16 pre_div;
+	u8 pre_div;
 	u16 m;
 	u16 n;
 };
